% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gamma.R
\name{make_gamma}
\alias{make_gamma}
\title{Aggregation of estimated error}
\usage{
make_gamma(E, type = "mu", ind = 1)
}
\arguments{
\item{E}{A matrix representing the estimated error matrix at some time t of dim (p,q).}

\item{type}{Character input, choice from one of 'mu', 'alpha' and 'beta'. Default is 'mu'.}

\item{ind}{integer denoting the index of interest, only used when type is 'alpha' or 'beta'. Default is 1.}
}
\value{
A numeric number
}
\description{
Computing the aggregated estimated error at some index for constructing asymptotic normality
}
