% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xy.R
\name{make_xy}
\alias{make_xy}
\title{Construction of series for testing MEFM}
\usage{
make_xy(E, type = "alpha")
}
\arguments{
\item{E}{An array representing the sequence of estimated error matrix of dim (T,p,q).}

\item{type}{Character input, either 'alpha' or 'beta'. Default is 'alpha'.}
}
\value{
A vector representing the constructed x or y series
}
\description{
Constructing x or y series for the MEFM testing
}
