% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qHat.R
\name{qHat}
\alias{qHat}
\title{Estimated theta quantile based on a given series}
\usage{
qHat(xt, theta = 0.95)
}
\arguments{
\item{xt}{A vector representing a series.}

\item{theta}{A value from 0 to 1. Default is 0.95.}
}
\value{
A numeric number
}
\description{
Computing the estimated quantile according to a given series for a given level of theta
}
